requirements.txt
aiogram==2.14.3
aiohttp==3.7.4.post0; python_version >= "3.6"
aioschedule==0.5.2
aiosqlite==0.16.1; python_version >= "3.7" and python_version < "4.0"
amqp==5.0.6; python_version >= "3.6"
async-timeout==3.0.1; python_full_version >= "3.5.3" and python_version >= "3.6"
asyncpg==0.24.0; python_full_version >= "3.6.0"
attrs==21.2.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6"
babel==2.9.1; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0"
billiard==3.6.4.0; python_version >= "3.6"
cachetools==4.2.2; python_version >= "3.6" and python_version < "4.0"
celery==5.1.2; python_version >= "3.6"
certifi==2021.5.30; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6"
chardet==4.0.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6"
charset-normalizer==2.0.4; python_full_version >= "3.6.0" and python_version >= "3.6"
click-didyoumean==0.0.3; python_version >= "3.6"
click-plugins==1.1.1; python_version >= "3.6"
click-repl==0.2.0; python_version >= "3.6"
click==7.1.2; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6"
google-api-core==2.0.0; python_version >= "3.6"
google-api-python-client==2.17.0; python_version >= "3.6"
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.5; python_version >= "3.6"
google-auth==2.0.1; python_version >= "3.6"
googleapis-common-protos==1.53.0; python_version >= "3.6"
httplib2==0.19.1; python_version >= "3.6"
idna==3.2; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6"
iso8601==0.1.16; python_version >= "3.7" and python_version < "4.0"
kombu==5.1.0; python_version >= "3.6"
multidict==5.1.0; python_version >= "3.6"
oauth2client==4.1.3
oauthlib==3.1.1; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
prompt-toolkit==3.0.20; python_full_version >= "3.6.2" and python_version >= "3.6"
protobuf==3.17.3; python_version >= "3.6"
pyasn1-modules==0.2.8; python_version >= "3.6"
pyasn1==0.4.8; python_version >= "3.5" and python_version < "4"
pyparsing==2.4.7; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.6"
pypika-tortoise==0.1.1; python_version >= "3.7" and python_version < "4.0"
pytz==2021.1
redis==3.5.3; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.5.0")
requests-oauthlib==1.3.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
requests==2.26.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6"
rsa==4.7.2; python_version >= "3.6" and python_version < "4"
six==1.16.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.6"
tortoise-orm==0.17.6; python_version >= "3.7" and python_version < "4.0"
typing-extensions==3.10.0.0; python_version >= "3.7" and python_version < "4.0"
uritemplate==3.0.1; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
urllib3==1.26.6; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version < "4" and python_version >= "3.6"
vine==5.0.0; python_version >= "3.6"
wcwidth==0.2.5; python_full_version >= "3.6.2" and python_version >= "3.6"
yarl==1.6.3; python_version >= "3.6"
